+++ src/main/java/com/ivanolmo/kanbantaskmanager/mapper/SubtaskMapper.java
@@ -1,7 +1,7 @@
 package com.ivanolmo.kanbantaskmanager.mapper;
 
 import com.ivanolmo.kanbantaskmanager.entity.Subtask;
-import com.ivanolmo.kanbantaskmanager.entity.dto.SubtaskDTO;
+import com.ivanolmo.kanbantaskmanager.dto.SubtaskDTO;
 import org.springframework.stereotype.Component;
 
 import java.util.Optional;
