+++ src/main/java/com/ivanolmo/kanbantaskmanager/service/ColumnService.java
@@ -1,6 +1,6 @@
 package com.ivanolmo.kanbantaskmanager.service;
 
-import com.ivanolmo.kanbantaskmanager.entity.dto.ColumnDTO;
+import com.ivanolmo.kanbantaskmanager.dto.ColumnDTO;
 
 public interface ColumnService {
   ColumnDTO addColumnToBoard(Long boardId, ColumnDTO columnDTO);
