+++ src/main/java/com/ivanolmo/kanbantaskmanager/service/SubtaskService.java
@@ -1,6 +1,6 @@
 package com.ivanolmo.kanbantaskmanager.service;
 
-import com.ivanolmo.kanbantaskmanager.entity.dto.SubtaskDTO;
+import com.ivanolmo.kanbantaskmanager.dto.SubtaskDTO;
 
 public interface SubtaskService {
   SubtaskDTO addSubtaskToTask(Long taskId, SubtaskDTO subtaskDTO);
